+++ scopesim/utils.py
@@ -15,7 +15,6 @@ from astropy import units as u
 from astropy.io import fits
 from astropy.io import ascii as ioascii
 from astropy.table import Column, Table
-from astropy.table.pprint import TableFormatter
 
 from . import rc
 
@@ -562,18 +561,6 @@ def convert_table_comments_to_dict(tbl):
         warnings.warn("No comments in table")
 
     return comments_dict
-
-
-def table_to_rst(tbl, indent=0):
-    tbl_fmtr = TableFormatter()
-    lines, outs = tbl_fmtr._pformat_table(tbl, max_width=-1, max_lines=-1)
-    lines[1] = lines[1].replace("-", "=")
-    lines = [lines[1]] + lines + [lines[1]]
-
-    indent = " " * indent
-    rst_str = indent + ("\n" + indent).join(lines)
-
-    return rst_str
 
 
 def change_table_entry(tbl, col_name, new_val, old_val=None, position=None):
